Docker working
Docker is working, large delay in running the docker container because it re-assembles the npm/node cache each time the container starts.  Leaving this is because of the development nature of the app at this stage.  Ant commands have been streamlined and updated to support the updated functionality.  Instructions and requirements added to the readme.md.