Implemented find results by teacher and post routes
Find results by teacher implemented similarly to finding results
by student.  Post for SINGULAR result implemented (will require
updating to handle the expected client test dump at completion of
a test).  'correct_answer' column changed to 'correctly_answered'
column of boolean type.  Post route does an eval on the supplied
operands and operation and compares to student answer to populate
this column.  Storing the correct answer to have to do an evalu-
ation later to determine whether the student was correct or not
seemed inefficient.  This change can be reverted if required.