Removed attempt_number field from result model
attempt_number field determined to be unnecessary as live
version will (likely) have timestamps enabled.  As testId was
explicitly declared within the model only to enforce a composite
uniqueness constraint with attempt_number, testId was removed as
well.  Test.hasMany(result) was added again to database.js in
order to generate the foreign key in absence of explicit declaration.