Test model updated to reflect new input validation requirements
baseNumber and operation fields added of type integer and of type
enum, respectively.  level field changed to type virtual and will
return a combination of baseNumber and operation; if attempted to
be set directly, an error will be thrown.  POST route will need
to be changed to set baseNumber and operation individually so that
they can be individually validated once validation is implemented.
The first GET route, beginning at line 36, has been updated with
field exclusion that will need to be replicated for all remaining
GET routes within tests.js.  Similar refactoring will be required
for the student model and students routes.