Resolved VIRTUAL level field displaying as 'undefinedundefined'
VIRTUAL fields generate only when the underlying seqelize data
set is converted to JSON payload.  Thus, to return the VIRTUAL
field only, the data set must be forced into a JSON string, then
parsed into a JSON object so that the field(s) the VIRTUAL field
is(/are) dependent on can be deleted without affecting the
VIRTUAL field itself.