Correct response format for routes with ID parameter
Return one object with ID parameter rather than an array with
only one object as its contents.  This is accomplished using
findOne rather than findByPk since other manipulations to the
data need to occur.  This is safe as since ID is the pk, only
one resut would be returned from the database in either case.