category field in TEST changed to practice to match API specification
Routes with id field as a parameter other than tests/:id were
commented out as all others are unnecessary.  This results from
id field being the primary key, so any and all routes with :id will
definitionally return one result, thus it is the only necessary
parameter in a request.  Also, id was removed from the POST route;
id is an auto-incremented primary key and should not be set
manually.